More consistent approach to reviewing ahead
Previously extending the reviewed passage didn't work properly when
reviewing ahead because we only looked for due lines when extending.
Now we use the same criterion for picking the line to review and for
extending. At the cost that we can only review ahead lines that are
due later today.