Complete rewrite to replace FSRS with simpler, more specific algorithm
It has become apparent that FSRS just isn't appropriate for this stuff
because it's designed for long-term retention. Due dates for lines that
have been well-learnt rapidly push out to several months, which isn't
useful for this purpose. We were also having to stick in hack the initial
learning process. FSRS did seem to work well for the intermediate case,
but I don't think the benefits are enough to justify all the hacks.